drivers/iio/frequency/cf_axi_dds: Add support for AD9162

+++ drivers/iio/frequency/ad9122.c
@@ -350,7 +350,7 @@ static int ad9122_get_clks(struct cf_axi_converter *conv)
 	return 0;
 }
 
-static unsigned long ad9122_get_data_clk(struct cf_axi_converter *conv)
+static unsigned long long ad9122_get_data_clk(struct cf_axi_converter *conv)
 {
 	return clk_get_rate(conv->clk[CLK_DATA]);
 }
